requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
psycopg2==2.9.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.0
PyMySQL==0.9.3
PyJWT==2.8.0
redis==5.0.1
Celery==5.3.4
flower==2.0.1
python-dotenv==1.0.0
